Added a learnability transparency parameter lamba. added continuous distance calculation in support of the lambda mechanism for continuous structures. Also experimentally implemented anticonformity.

+++ src/Graph.hpp
@@ -3,7 +3,7 @@
 
 #include "Types.hpp" 
 
-std::vector<int> computeDistances(const AdjacencyMatrix& adjMatrix, Trait root);
+std::vector<double> computeDistances(const AdjacencyMatrix& adjMatrix, Trait root);
 
 std::vector<Trait> parentTraits(const AdjacencyMatrix& adjMatrix, Trait trait);
